remove the command from the scripts package
