+++ bin/changelog/cli.js
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 'use strict';
 
-const { makeChangeLog } = require( '../utils' );
+const { makeChangeLog } = require( './' );
 
 makeChangeLog();
