+++ package.json
@@ -79,6 +79,7 @@
 		"@babel/runtime-corejs3": "7.9.2",
 		"@babel/traverse": "7.9.0",
 		"@jest/types": "25.3.0",
+		"@octokit/graphql": "4.3.1",
 		"@octokit/rest": "16.26.0",
 		"@octokit/webhooks": "7.1.0",
 		"@storybook/addon-a11y": "5.3.2",
@@ -191,7 +192,7 @@
 		"build:packages": "npm run build:package-types && node ./bin/packages/build.js",
 		"build:package-types": "node ./bin/packages/validate-typescript-version.js && tsc --build",
 		"build": "npm run build:packages && wp-scripts build",
-		"changelog": "wp-scripts changelog",
+		"changelog": "./bin/changelog/cli.js",
 		"check-engines": "wp-scripts check-engines",
 		"check-licenses": "concurrently \"wp-scripts check-licenses --prod --gpl2\" \"wp-scripts check-licenses --dev\"",
 		"precheck-local-changes": "npm run docs:build",
