+++ packages/scripts/package.json
@@ -33,7 +33,6 @@
 		"wp-scripts": "./bin/wp-scripts.js"
 	},
 	"dependencies": {
-		"@octokit/graphql": "4.3.1",
 		"@svgr/webpack": "^5.2.0",
 		"@wordpress/babel-preset-default": "file:../babel-preset-default",
 		"@wordpress/dependency-extraction-webpack-plugin": "file:../dependency-extraction-webpack-plugin",
@@ -61,7 +60,6 @@
 		"minimist": "^1.2.0",
 		"npm-package-json-lint": "^5.0.0",
 		"prettier": "npm:wp-prettier@1.19.1",
-		"promptly": "3.0.3",
 		"puppeteer": "npm:puppeteer-core@3.0.0",
 		"read-pkg-up": "^1.0.1",
 		"resolve-bin": "^0.4.0",
