+++ packages/scripts/utils/index.js
@@ -23,7 +23,6 @@ const {
 } = require( './env' );
 const { fromProjectRoot, fromConfigRoot, hasProjectFile } = require( './file' );
 const { hasPackageProp } = require( './package' );
-const { makeChangeLog } = require( './changelog/make-change-log' );
 
 module.exports = {
 	buildWordPress,
@@ -44,5 +43,4 @@ module.exports = {
 	downloadWordPressZip,
 	mergeYAMLConfigs,
 	spawnScript,
-	makeChangeLog,
 };
